DisplayField型を別ファイルに移動
