+++ household-account-book/app/_calculate/asMonthly.tsx
@@ -1,5 +1,5 @@
 import calculateDisplayField from "./devide";
-import { DisplayField } from "./types";
+import { DisplayField } from "./devide";
 import { TableRecord } from "../_storage/handler";
 
 function asMonthly(dataTable: TableRecord[]): DisplayField[] {
