--- household-account-book/app/_calculate/types.tsx
@@ -1,8 +1,0 @@
-type DisplayField = {
-    id: string;
-    title: string;
-    amount: number;
-    createdAt: Date;
-}
-
-export type { DisplayField };
